fix: break the main file into separate files
